- call mke2fs with -I 128 (bnc #358483)
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@1252 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ Makefile
@@ -76,7 +76,7 @@ install:
 	cp -a doc/schema/   ${PACKDOCVZ}
 	cp -a doc/*.pdf     ${PACKDOCVZ}
 	cp -a doc/COPYING   ${PACKDOCVZ}
-	cp -a doc/ChangeLog ${PACKDOCVZ}
+	test -e doc/ChangeLog && cp -a doc/ChangeLog ${PACKDOCVZ} || true
 
 	#============================================
 	# kiwi manual pages
